+++ axelor-account/src/main/java/com/axelor/apps/account/service/batch/AccountingBatchViewService.java
@@ -27,4 +27,6 @@ public interface AccountingBatchViewService {
   List<Account> getClosureAccountSet(AccountingBatch accountingBatch, boolean closeAllAccounts);
 
   List<Account> getOpeningAccountSet(AccountingBatch accountingBatch, boolean openAllAccounts);
+
+  Long countAllAvailableAccounts(AccountingBatch accountingBatch, boolean isClosure);
 }