+++ axelor-account/src/main/java/com/axelor/apps/account/service/moveline/MoveLineGroupServiceImpl.java
@@ -160,6 +160,7 @@ public class MoveLineGroupServiceImpl implements MoveLineGroupService {
     moveLineAttrsService.addTaxLineRequired(move, moveLine, attrsMap);
     moveLineAttrsService.addCutOffPanelHidden(move, moveLine, attrsMap);
     moveLineAttrsService.addCutOffDatesRequired(move, moveLine, attrsMap);
+    moveLineAttrsService.addVatSystemSelectReadonly(move, moveLine, attrsMap);
 
     return attrsMap;
   }
@@ -204,6 +205,7 @@ public class MoveLineGroupServiceImpl implements MoveLineGroupService {
       moveLineAttrsService.addTaxLineRequired(move, moveLine, attrsMap);
       moveLineAttrsService.addCutOffPanelHidden(move, moveLine, attrsMap);
       moveLineAttrsService.addCutOffDatesRequired(move, moveLine, attrsMap);
+      moveLineAttrsService.addVatSystemSelectReadonly(move, moveLine, attrsMap);
     }
 
     return attrsMap;
@@ -230,6 +232,7 @@ public class MoveLineGroupServiceImpl implements MoveLineGroupService {
       moveLineAttrsService.addTaxLineRequired(move, moveLine, attrsMap);
       moveLineAttrsService.addCutOffPanelHidden(move, moveLine, attrsMap);
       moveLineAttrsService.addCutOffDatesRequired(move, moveLine, attrsMap);
+      moveLineAttrsService.addVatSystemSelectReadonly(move, moveLine, attrsMap);
     }
 
     return attrsMap;
@@ -344,6 +347,7 @@ public class MoveLineGroupServiceImpl implements MoveLineGroupService {
     analyticAttrsService.addAnalyticAxisAttrs(move.getCompany(), null, attrsMap);
     moveLineAttrsService.changeFocus(move, moveLine, attrsMap);
     moveLineAttrsService.addTaxLineRequired(move, moveLine, attrsMap);
+    moveLineAttrsService.addVatSystemSelectReadonly(move, moveLine, attrsMap);
 
     return attrsMap;
   }