+++ axelor-base/src/main/java/com/axelor/apps/base/service/PartnerPriceListDomainService.java
@@ -1,0 +1,9 @@
+package com.axelor.apps.base.service;
+
+import com.axelor.apps.base.db.Partner;
+
+public interface PartnerPriceListDomainService {
+  String getSalePartnerPriceListDomain(Partner partner);
+
+  String getPurchasePartnerPriceListDomain(Partner partner);
+}