+++ axelor-contract/src/main/java/com/axelor/apps/contract/batch/BatchContractInvoicing.java
@@ -18,6 +18,7 @@
  */
 package com.axelor.apps.contract.batch;
 
+import com.axelor.apps.account.db.Invoice;
 import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.service.exception.TraceBackService;
 import com.axelor.apps.contract.db.Contract;
@@ -98,14 +99,15 @@ public class BatchContractInvoicing extends BatchStrategy {
   }
 
   @Transactional
-  public void invoiceContracts(List<Contract> contractList) throws AxelorException {
+  public Invoice invoiceContracts(List<Contract> contractList) throws AxelorException {
     if (contractList.size() == 1) {
-      contractInvoicingService.invoicingContract(contractList.get(0));
+      return contractInvoicingService.invoicingContract(contractList.get(0));
     }
 
     if (contractList.size() > 1) {
-      contractInvoicingService.invoicingContracts(contractList);
+      return contractInvoicingService.invoicingContracts(contractList);
     }
+    return null;
   }
 
   @Override
@@ -113,17 +115,37 @@ public class BatchContractInvoicing extends BatchStrategy {
     int offset = 0;
     for (List<Long> idList : getIdsGroupedBy()) {
       ++offset;
+      List<Contract> contractsList = findContractsInList(idList);
       try {
-        invoiceContracts(findContractsInList(idList));
-        incrementDone();
+        Invoice invoice = invoiceContracts(contractsList);
+        if (invoice != null) {
+          invoice.addBatchSetItem(batchRepo.find(batch.getId()));
+          incrementDone(contractsList);
+        }
       } catch (Exception e) {
-        incrementAnomaly();
+        incrementAnomaly(contractsList);
         TraceBackService.trace(e, "Contract invoicing batch", batch.getId());
       }
       if (offset % getFetchLimit() == 0) {
         JPA.clear();
         findBatch();
       }
+    }
+  }
+
+  protected void incrementDone(List<Contract> contractsList) {
+    for (Contract contract : contractsList) {
+      contract.addBatchSetItem(batch);
+      super.incrementDone();
+    }
+  }
+
+  protected void incrementAnomaly(List<Contract> contractsList) {
+    findBatch();
+    for (Contract contract : contractsList) {
+      contract = contractRepository.find(contract.getId());
+      contract.addBatchSetItem(batch);
+      super.incrementAnomaly();
     }
   }