+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/saleorder/SaleOrderPurchaseServiceImpl.java
@@ -20,6 +20,7 @@ package com.axelor.apps.supplychain.service.saleorder;
 
 import com.axelor.apps.account.db.repo.AccountConfigRepository;
 import com.axelor.apps.base.AxelorException;
+import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.Partner;
 import com.axelor.apps.base.db.repo.PriceListRepository;
 import com.axelor.apps.base.db.repo.ProductRepository;
@@ -34,9 +35,11 @@ import com.axelor.apps.purchase.service.config.PurchaseConfigService;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.apps.supplychain.exception.SupplychainExceptionMessage;
+import com.axelor.apps.supplychain.service.IntercoService;
 import com.axelor.apps.supplychain.service.PurchaseOrderCreateSupplychainService;
 import com.axelor.apps.supplychain.service.PurchaseOrderLineServiceSupplyChain;
 import com.axelor.apps.supplychain.service.PurchaseOrderSupplychainService;
+import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.auth.AuthUtils;
 import com.axelor.i18n.I18n;
 import com.google.inject.Inject;
@@ -64,6 +67,8 @@ public class SaleOrderPurchaseServiceImpl implements SaleOrderPurchaseService {
   protected AppBaseService appBaseService;
   protected PartnerPriceListService partnerPriceListService;
   protected SupplierCatalogService supplierCatalogService;
+  protected AppSupplychainService appSupplychainService;
+  protected IntercoService intercoService;
 
   @Inject
   public SaleOrderPurchaseServiceImpl(
@@ -75,7 +80,9 @@ public class SaleOrderPurchaseServiceImpl implements SaleOrderPurchaseService {
       PurchaseConfigService purchaseConfigService,
       AppBaseService appBaseService,
       PartnerPriceListService partnerPriceListService,
-      SupplierCatalogService supplierCatalogService) {
+      SupplierCatalogService supplierCatalogService,
+      AppSupplychainService appSupplychainService,
+      IntercoService intercoService) {
     this.purchaseOrderSupplychainService = purchaseOrderSupplychainService;
     this.purchaseOrderCreateSupplychainService = purchaseOrderCreateSupplychainService;
     this.purchaseOrderLineServiceSupplychain = purchaseOrderLineServiceSupplychain;
@@ -85,6 +92,8 @@ public class SaleOrderPurchaseServiceImpl implements SaleOrderPurchaseService {
     this.appBaseService = appBaseService;
     this.partnerPriceListService = partnerPriceListService;
     this.supplierCatalogService = supplierCatalogService;
+    this.appSupplychainService = appSupplychainService;
+    this.intercoService = intercoService;
   }
 
   @Override
@@ -191,6 +200,12 @@ public class SaleOrderPurchaseServiceImpl implements SaleOrderPurchaseService {
       purchaseOrder.setInAti(true);
     } else {
       purchaseOrder.setInAti(false);
+    }
+
+    if (appSupplychainService.getAppSupplychain().getIntercoFromPurchase()) {
+      Company intercoCompany = intercoService.findIntercoCompany(supplierPartner);
+      boolean isInterco = intercoCompany != null;
+      purchaseOrder.setInterco(isInterco);
     }
 
     purchaseOrder.setNotes(supplierPartner.getPurchaseOrderComments());