+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/saleorder/SaleOrderServiceSupplychainImpl.java
@@ -52,6 +52,7 @@ import com.axelor.apps.supplychain.db.Timetable;
 import com.axelor.apps.supplychain.exception.SupplychainExceptionMessage;
 import com.axelor.apps.supplychain.service.AccountingSituationSupplychainService;
 import com.axelor.apps.supplychain.service.PartnerLinkSupplychainService;
+import com.axelor.apps.supplychain.service.SaleInvoicingStateService;
 import com.axelor.apps.supplychain.service.TrackingNumberSupplychainService;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.i18n.I18n;
@@ -77,6 +78,7 @@ public class SaleOrderServiceSupplychainImpl extends SaleOrderServiceImpl
   protected TrackingNumberSupplychainService trackingNumberSupplychainService;
 
   protected PartnerLinkSupplychainService partnerLinkSupplychainService;
+  protected SaleInvoicingStateService saleInvoicingStateService;
 
   @Inject
   public SaleOrderServiceSupplychainImpl(
@@ -95,7 +97,8 @@ public class SaleOrderServiceSupplychainImpl extends SaleOrderServiceImpl
       SaleOrderStockService saleOrderStockService,
       AccountingSituationSupplychainService accountingSituationSupplychainService,
       TrackingNumberSupplychainService trackingNumberSupplychainService,
-      PartnerLinkSupplychainService partnerLinkSupplychainService) {
+      PartnerLinkSupplychainService partnerLinkSupplychainService,
+      SaleInvoicingStateService saleInvoicingStateService) {
     super(
         appBaseService,
         saleOrderLineRepo,
@@ -113,6 +116,7 @@ public class SaleOrderServiceSupplychainImpl extends SaleOrderServiceImpl
     this.accountingSituationSupplychainService = accountingSituationSupplychainService;
     this.trackingNumberSupplychainService = trackingNumberSupplychainService;
     this.partnerLinkSupplychainService = partnerLinkSupplychainService;
+    this.saleInvoicingStateService = saleInvoicingStateService;
   }
 
   public SaleOrder getClientInformations(SaleOrder saleOrder) {
@@ -270,6 +274,7 @@ public class SaleOrderServiceSupplychainImpl extends SaleOrderServiceImpl
     }
 
     saleOrderStockService.fullyUpdateDeliveryState(saleOrder);
+    saleInvoicingStateService.updateInvoicingState(saleOrder);
     saleOrder.setOrderBeingEdited(false);
 
     if (appSupplychainService.getAppSupplychain().getCustomerStockMoveGenerationAuto()) {